issue-2751 use new ThreadHandler to replace ThreadHelper
